move join_one.rs
