+++ enclone_core/Cargo.toml
@@ -30,6 +30,8 @@ regex = { version = "1.3.1", default-features = false, features = ["std", "perf"
 serde = "1.0.90"
 serde_derive = "1.0.102"
 serde_json = "*"
+stats_utils = "0.1.2"
+stirling_numbers = "0.1.2"
 string_utils = "0.1.1"
 superslice = "1"
 vdj_ann = { git = "https://github.com/10XGenomics/rust-toolbox.git", rev="0fc3f83b280ed4a11de489b4fdbc755f8cd53783" }
