+++ enclone_core/src/lib.rs
@@ -1,6 +1,7 @@
 // Copyright (c) 2021 10x Genomics, Inc. All rights reserved.
 
 pub mod defs;
+pub mod join_one;
 pub mod mammalian_fixed_len;
 pub mod print_tools;
 pub mod testlist;
